+++ src/db.mli
@@ -51,7 +51,7 @@ val refresh_library : library:string -> (Path.rel * Doc.t) list
 
 val resolve_missing_files : library:string -> Path.rel list
 
-val find_duplicates : unit -> (string * Path.rel) list
+val find_duplicates : unit -> ((string * Path.rel) list) list
   
 val flush_metadata : unit -> unit
 val flush_libconfig : unit -> unit