+++ src/font.mli
@@ -18,3 +18,5 @@
 (******************************************************************************)
 
 val calc_font_height : widget:GObj.widget -> ?ypad:int -> int -> int
+
+val pango_quote : string -> string